CSL: Newer releases of g++ had further enhanced their optimization capability, taking advantages of more places where signe dinteger arithmetic on initially positive values overflowed into the carry bit is officially undefined - and so they are allowed to rely on it never happening and save code. However in CSL there was at least one new place (in the code to print negative bignums) where this overflw could be triggered and where the new g++ optimizations meanst that the code printed bad results when compiled with optimization enabled - but not when build for debugging! I have made some attempt to look for other places where the same might arise and put in some ugly casts to uint32_t (because arithmetic operations on unsigned integers have to behave "the way and old fashioned C programmer would expect". At present I do not know a systematic way to make the code safe against such effects -- and even more certainly I do not know a neat and natural way. However it makes me glad that I am working on replacing the old CSL bignum code - because that code worked with 31-bit digits in 32-bit words precisely so that addition would overflow into the sign bit where it would be easy to notice!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5023 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith.h
@@ -96,7 +96,7 @@ extern unsigned char msd_table[256], lsd_table[256];
 #define top_bit_set(n)     (((int32_t)(n)) < 0)
 #define top_bit(n)         ((int32_t)(((uint32_t)(n)) >> 31))
 #define set_top_bit(n)     ((int32_t)((uint32_t)(n) | (uint32_t)0x80000000U))
-#define clear_top_bit(n)   ((int32_t)(n) & 0x7fffffff)
+#define clear_top_bit(n)   ((int32_t)((uint32_t)(n) & 0x7fffffff))
 
 // As with fixnum_of_int I need to take care here that the arithmetic I do
 // here can not overflow, since if it did the behaviour would be undefined
